moved 2 methods
