+++ client/src/pages/FAQPage.js
@@ -1,4 +1,3 @@
-import fetch, { Body } from 'node-fetch';
 import React from 'react';
 import "./css/faqpage.css";
 
@@ -7,45 +6,7 @@ export default class FAQPage extends React.Component {
     constructor(props){
         super(props);
         this.printQuestions = this.printQuestions.bind(this);
-
-        this.updateFAQ = this.updateFAQ.bind(this)
-        this.addFAQ = this.addFAQ.bind(this)
     }
-
-
-  
-
-      updateFAQ = () => {
-        let req = { method: "PUT", 
-        headers: { "Content-Type": "application/json" },
-        body: JSON.stringify({id: 1 , question: "hej", answer: "coolt"})
-
-        }
-        fetch("/api/faq",req)
-          .then((response) => response.text())
-          .then((response) => {
-            console.log(response);
-          });
-      }
-
-
-      addFAQ = () => {
-        let req = { method: "POST", 
-        headers: { "Content-Type": "application/json" },
-        body: JSON.stringify({question: "New q", answer: "new a"})
-
-        }
-        fetch("/api/faq",req)
-          .then((response) => response.text())
-          .then((response) => {
-            console.log(response);
-          });
-      }
-
-
-      
-
-   
 
     printQuestions = () => {
             let arr = [];
